+++ jpress-core/src/main/java/io/jpress/JPressInitializer.java
@@ -31,7 +31,7 @@ import io.jpress.core.addon.handler.AddonHandlerProcesser;
 import io.jpress.core.addon.interceptor.AddonInterceptorProcesser;
 import io.jpress.core.install.InstallHandler;
 import io.jpress.core.menu.MenuManager;
-import io.jpress.core.support.EhcacheSupporter;
+import io.jpress.core.support.ehcache.EhcacheSupporter;
 import io.jpress.core.wechat.WechatAddonManager;
 import io.jpress.web.captcha.JPressCaptchaCache;
 import io.jpress.web.handler.JPressHandler;
