--- jpress-core/src/main/java/io/jpress/core/ext/field/renders/DatetimeRender.java
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) 2016-2019, Michael Yang 杨福海 (fuhai999@gmail.com).
- * <p>
- * Licensed under the GNU Lesser General Public License (LGPL) ,Version 3.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.gnu.org/licenses/lgpl-3.0.txt
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.jpress.core.ext.field.renders;
-
-import io.jpress.core.ext.field.SmartField;
-import io.jpress.core.ext.field.SmartFieldRender;
-
-/**
- * @author Michael Yang 杨福海 （fuhai999@gmail.com）
- * @version V1.0
- * @Title: Datetime 时间选择器的渲染器
- */
-public class DatetimeRender implements SmartFieldRender {
-
-
-    @Override
-    public String onRender(SmartField field, Object value) {
-        throw new IllegalStateException("DatetimeRender not finished.");
-    }
-
-}
