+++ jpress-core/src/main/java/io/jpress/core/support/smartfield/renders/FileRender.java
@@ -13,33 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.jpress.core.ext.field.renders;
+package io.jpress.core.support.smartfield.renders;
 
-import io.jpress.core.ext.field.SmartField;
-import io.jpress.core.ext.field.SmartFieldRender;
+import io.jpress.core.support.smartfield.SmartField;
+import io.jpress.core.support.smartfield.SmartFieldRender;
 
 /**
  * @author Michael Yang 杨福海 （fuhai999@gmail.com）
  * @version V1.0
  * @Title: Input 输入框的渲染器
  */
-public class InputRender implements SmartFieldRender {
-
-    protected static String template = "" +
-            "<div class=\"form-group\">\n" +
-            "    <label class=\"col-sm-2 control-label\">{label}</label>\n" +
-            "    <div class=\"col-sm-6\">\n" +
-            "        <input type=\"text\" class=\"form-control\"\n" +
-            "               id=\"{id}\"\n" +
-            "               name=\"{name}\"\n" +
-            "               placeholder=\"{placeholder}\" value=\"{value}\" {attrs}/>\n" +
-            "        <p class=\"help-block\">{helpText}</p>\n" +
-            "    </div>\n" +
-            "</div>";
+public class FileRender implements SmartFieldRender {
 
     @Override
     public String onRender(SmartField field, Object value) {
-        return RenderKit.render(template, field, value);
+        throw new IllegalStateException("FileRender not finished.");
     }
 
 }
