+++ jpress-core/src/main/java/io/jpress/core/support/smartfield/renders/InputRender.java
@@ -13,36 +13,33 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.jpress.core.ext.field.renders;
+package io.jpress.core.support.smartfield.renders;
 
-import io.jpress.core.ext.field.SmartField;
-import io.jpress.core.ext.field.SmartFieldRender;
-
-import java.util.Objects;
+import io.jpress.core.support.smartfield.SmartField;
+import io.jpress.core.support.smartfield.SmartFieldRender;
 
 /**
  * @author Michael Yang 杨福海 （fuhai999@gmail.com）
  * @version V1.0
- * @Title: Switch 的渲染器
+ * @Title: Input 输入框的渲染器
  */
-public class SwitchRender implements SmartFieldRender {
+public class InputRender implements SmartFieldRender {
 
     protected static String template = "" +
             "<div class=\"form-group\">\n" +
             "    <label class=\"col-sm-2 control-label\">{label}</label>\n" +
             "    <div class=\"col-sm-6\">\n" +
-            "        <input type=\"checkbox\" {checked} class=\"switchery\"\n" +
-            "               data-for=\"{id}\" value=\"true\">\n" +
+            "        <input type=\"text\" class=\"form-control\"\n" +
+            "               id=\"{id}\"\n" +
+            "               name=\"{name}\"\n" +
+            "               placeholder=\"{placeholder}\" value=\"{value}\" {attrs}/>\n" +
             "        <p class=\"help-block\">{helpText}</p>\n" +
-            "        <input type=\"hidden\" id=\"{id}\" name=\"{name}\">\n" +
             "    </div>\n" +
             "</div>";
 
     @Override
     public String onRender(SmartField field, Object value) {
-        String checked = (value == null && "true".equals(field.getValue()))
-                ? "checked"
-                : (Objects.equals("true", value) ? "checked" : "");
-        return RenderKit.render(template, field, value).replace("{checked}", checked);
+        return RenderKit.render(template, field, value);
     }
+
 }
