+++ jpress-core/src/main/java/io/jpress/core/support/smartfield/renders/RenderKit.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.jpress.core.ext.field.renders;
+package io.jpress.core.support.smartfield.renders;
 
 import io.jboot.utils.StrUtil;
-import io.jpress.core.ext.field.SmartField;
+import io.jpress.core.support.smartfield.SmartField;
 
 
 public class RenderKit {
